requirement.txt
openai>=1.40.0
python-dotenv>=1.0.1
gradio>=4.0.0
pandas>=2.0.0
matplotlib>=3.8.0
chromadb>=0.5.0
llama-index-core>=0.11.0
llama-index-readers-file>=0.2.0
llama-index-vector-stores-chroma>=0.2.0
duckduckgo-search>=6.2.0
